pyproject.toml
[project]
name = "deepsse"
version = "0.1.0"
description = "Deep Learing based DOA estimator for variable and unkown number of signals"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "h5py>=3.12.1",
  "joblib>=1.4.2",
  "matplotlib>=3.9.2",
  "munkres>=1.1.4",
  "numpy>=2.1.2",
  "pandas>=2.2.3",
  "ruamel-yaml>=0.18.6",
  "scikit-learn>=1.5.2",
  "tensorboard>=2.18.0",
  "torch>=2.4.1",
  "tqdm>=4.66.5",
]
